requirements.txt
openai>=1.0.0
watchdog>=3.0.0
pillow>=10.0.0
python-dotenv>=1.0.0